scripts: Reference download script
This is an example script that can be used to fetch all of
the Cloudwatch logs of the `us-east-1` region into a fixtures
directory that can be used to populate the fake cloudwatch
server with.

The fixtures directory is in .gitignore because this is your
production data and you do not want to commit it into git.